// src/types/auth.types.ts
export interface AuthUser {
  id: string
  username: string
  email: string
}

export interface LoginResponse {
  token: string
  id: string
  username: string
  email: string
}

export interface LoginPayload {
  email: string
  password: string
}

export interface RegisterPayload {
  username: string
  email: string
  password: string
}